upload a006 (C++)

+++ a006/a006.cpp
@@ -1,4 +1,5 @@
 #include <iostream>
+#include <cmath>
 using namespace std;
 
 int main() {
@@ -9,11 +10,11 @@ int main() {
     int d = b * b - 4 * a *c;
 
     if (d > 0) {
-        int root1 = (-b + sqrt(d)) / (2 * a);
-        int root2 = (-b - sqrt(d)) / (2 * a);
+        int root1 = int(-b + sqrt(d)) / (2 * a);
+        int root2 = int(-b - sqrt(d)) / (2 * a);
         cout << "Two different roots x1=" << root1 << " , x2=" << root2;
     } else if (d == 0) {
-        int root = (-b + sqrt(d)) / (2 * a);
+        int root = int(-b + sqrt(d)) / (2 * a);
         cout << "Two same roots x=" << root;
     } else {
         cout << "No real root";
